+++ src/main/java/com/diffbydevs/velog_clone/follow/entity/Follow.java
@@ -1,7 +1,7 @@
 package com.diffbydevs.velog_clone.follow.entity;
 
 import com.diffbydevs.velog_clone.common.entity.BaseEntity;
-import com.diffbydevs.velog_clone.user.entity.User;
+import com.diffbydevs.velog_clone.user.repository.entity.User;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
